# openapi.yaml
openapi: 3.0.3
info:
  title: MindWeaver API
  description: |
    MindWeaver 个人知识/研究动态监控工具 API

    功能特性：
    - RSS/Atom 订阅源管理
    - 条目获取、去重、存储和检索
    - 分类管理
    - 过滤规则管理
    - 自动调度获取
    - AI 摘要生成
    - 邮件摘要推送

  version: 1.0.0
  contact:
    name: MindWeaver
servers:
  - url: http://localhost:8000
    description: 本地开发服务器
paths:
...
  /api/dashboard/activity:
    get:
      tags: [System]
      summary: 获取最近活动
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: 返回条数
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Entry'
...
  /api/system/cleanup:
    post:
      tags: [System]
      summary: 清理旧条目
      description: 删除指定天数之前的条目
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                days:
                  type: integer
                  default: 90
                  description: 保留最近多少天的条目
      responses:
        '200':
          description: 清理完成
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      deleted_count:
                        type: integer
                  message:
                    type: string
  /api/system/export/entries:
    get:
      tags: [System]
      summary: 导出条目
      description: 导出条目为 JSON 文件
      parameters:
        - name: feed_id
          in: query
          schema:
            type: integer
          description: 按订阅源筛选
        - name: limit
          in: query
          schema:
            type: integer
            default: 1000
          description: 导出数量限制
      responses:
        '200':
          description: JSON 文件下载
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Entry'
          headers:
            Content-Disposition:
              schema:
                type: string
              description: attachment; filename=entries.json
...
  /api/system/config/llm:
    put:
      tags: [System]
      summary: 更新 LLM 配置
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LLMConfig'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/system/config/llm/test:
    post:
      tags: [System]
      summary: 测试 LLM 连接
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - provider
                - api_key
              properties:
                provider:
                  type: string
                  enum: [openai, anthropic, deepseek]
                  description: LLM 提供商
                api_key:
                  type: string
                  description: API 密钥
                api_base:
                  type: string
                  description: API 基础 URL（可选）
                model:
                  type: string
                  description: 模型名称（可选）
      responses:
        '200':
          description: 连接成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: 连接失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/system/config/email:
    put:
      tags: [System]
      summary: 更新邮件配置
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailConfig'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/system/config/email/test:
    post:
      tags: [System]
      summary: 测试邮件连接
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - smtp_host
                - smtp_port
                - username
                - password
              properties:
                smtp_host:
                  type: string
                smtp_port:
                  type: integer
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: 连接成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: 连接失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/system/config/digest:
    put:
      tags: [System]
      summary: 更新摘要配置
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DigestConfig'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: 操作成功
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: 错误信息
        status:
          type: integer
          example: 400
...
    Entry:
      type: object
      properties:
        id:
          type: integer
        feed_id:
          type: integer
          description: 所属订阅源 ID
        title:
          type: string
          description: 条目标题
        link:
          type: string
          format: uri
          description: 原始链接
        author:
          type: string
          description: 作者
        summary:
          type: string
          description: AI 生成的摘要
        content:
          type: string
          description: 完整内容
        published_at:
          type: string
          format: date-time
          description: 发布时间
        fetched_at:
          type: string
          format: date-time
          description: 获取时间
        tags:
          type: array
          items:
            type: string
          description: 标签列表
        language:
          type: string
          description: 语言代码
        reading_time_seconds:
          type: integer
          description: 预计阅读时间（秒）
        title_hash:
          type: string
          description: 标题哈希（用于去重）
        link_hash:
          type: string
          description: 链接哈希（用于去重）
        content_hash:
          type: string
          description: 内容哈希（用于去重）
        enabled:
          type: boolean
          description: 是否启用（未读=true，已读=false）
          default: true
...
    LLMConfig:
      type: object
      properties:
        provider:
          type: string
          enum: [openai, anthropic, deepseek]
        api_key:
          type: string
        api_base:
          type: string
        model:
          type: string
        temperature:
          type: number
          minimum: 0
          maximum: 2
        max_tokens:
          type: integer
    EmailConfig:
      type: object
      properties:
        smtp_host:
          type: string
        smtp_port:
          type: integer
        username:
          type: string
        password:
          type: string
        use_tls:
          type: boolean
          default: true
        from_addr:
          type: string
          format: email
        to_addrs:
          type: array
          items:
            type: string
            format: email
    DigestConfig:
      type: object
      properties:
        enabled:
          type: boolean
          default: false
        schedule:
          type: string
          description: Cron 表达式
        max_entries:
          type: integer
          description: 每封邮件最大条目数
          default: 20
        include_summary:
          type: boolean
          description: 是否包含 AI 摘要
          default: true